Added logging in to swagger
Also extracted some configuration to extension methods
